#27 Move Config structure outside of HCaptcha class
