+++ Example/HCaptcha_Tests/Core/HCaptcha__Config__Tests.swift
@@ -15,19 +15,84 @@ class HCaptcha__Config__Tests: XCTestCase {
         + "&assethost=https%3A%2F%2Fnewassets.hcaptcha.com&imghost=https%3A%2F%2Fimgs.hcaptcha.com"
         + "&reportapi=https%3A%2F%2Faccounts.hcaptcha.com"
 
-    var config: HCaptcha.Config?
-
     func createConfig(apiKey: String = "some-api-key",
                       host: String? = nil,
-                      customTheme: String? = nil) -> HCaptcha.Config? {
-        return try? HCaptcha.Config(apiKey: apiKey,
-                                    infoPlistKey: nil,
-                                    baseURL: URL(string: "https://localhost")!,
-                                    infoPlistURL: nil,
-                                    host: host,
-                                    customTheme: customTheme)
+                      customTheme: String? = nil) -> HCaptchaConfig? {
+        return try? HCaptchaConfig(apiKey: apiKey,
+                                   infoPlistKey: nil,
+                                   baseURL: URL(string: "https://localhost")!,
+                                   infoPlistURL: nil,
+                                   host: host,
+                                   customTheme: customTheme)
     }
 
+    func test__Base_URL() {
+        // Ensures baseURL failure when nil
+        do {
+            _ = try HCaptchaConfig(apiKey: "", infoPlistKey: nil, baseURL: nil, infoPlistURL: nil)
+            XCTFail("Should have failed")
+        } catch let e as HCaptchaError {
+            print(e)
+            XCTAssertEqual(e, HCaptchaError.baseURLNotFound)
+        } catch let e {
+            XCTFail("Unexpected error: \(e)")
+        }
+
+        // Ensures plist url if nil key
+        let plistURL = URL(string: "https://bar")!
+        let config1 = try? HCaptchaConfig(apiKey: "", infoPlistKey: nil, baseURL: nil, infoPlistURL: plistURL)
+        XCTAssertEqual(config1?.baseURL, plistURL)
+
+        // Ensures preference of given url over plist entry
+        let url = URL(string: "ftp://foo")!
+        let config2 = try? HCaptchaConfig(apiKey: "", infoPlistKey: nil, baseURL: url, infoPlistURL: plistURL)
+        XCTAssertEqual(config2?.baseURL, url)
+    }
+
+    func test__Base_URL_Without_Scheme() {
+        // Ignores URL with scheme
+        let goodURL = URL(string: "https://foo.bar")!
+        let config0 = try? HCaptchaConfig(apiKey: "", infoPlistKey: nil, baseURL: goodURL, infoPlistURL: nil)
+        XCTAssertEqual(config0?.baseURL, goodURL)
+
+        // Fixes URL without scheme
+        let badURL = URL(string: "foo")!
+        let config = try? HCaptchaConfig(apiKey: "", infoPlistKey: nil, baseURL: badURL, infoPlistURL: nil)
+        XCTAssertEqual(config?.baseURL.absoluteString, "http://" + badURL.absoluteString)
+    }
+
+    func test__API_Key() {
+        // Ensures key failure when nil
+        do {
+            _ = try HCaptchaConfig(apiKey: nil, infoPlistKey: nil, baseURL: nil, infoPlistURL: nil)
+            XCTFail("Should have failed")
+        } catch let e as HCaptchaError {
+            print(e)
+            XCTAssertEqual(e, HCaptchaError.apiKeyNotFound)
+        } catch let e {
+            XCTFail("Unexpected error: \(e)")
+        }
+
+        // Ensures plist key if nil key
+        let plistKey = "bar"
+        let config1 = try? HCaptchaConfig(
+            apiKey: nil,
+            infoPlistKey: plistKey,
+            baseURL: URL(string: "foo"),
+            infoPlistURL: nil
+        )
+        XCTAssertEqual(config1?.apiKey, plistKey)
+
+        // Ensures preference of given key over plist entry
+        let key = "foo"
+        let config2 = try? HCaptchaConfig(
+            apiKey: key,
+            infoPlistKey: plistKey,
+            baseURL: URL(string: "foo"),
+            infoPlistURL: nil
+        )
+        XCTAssertEqual(config2?.apiKey, key)
+    }
 
     func test__Locale__Nil() {
         let config = createConfig()
